+++ packages/origin-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @originloom/react
 
+## 0.7.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @originloom/shared@0.7.19
+
 ## 0.7.18
 
 ### Patch Changes